requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.5
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
tqdm==4.54.1
Werkzeug==1.0.1
timeloop
termcolor
prettytable
wget